requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
uvicorn==0.29.0
fastapi==0.110.1
instagrapi==1.1.4